# Backend/shopq_backend/public/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: ShopQ API
  description: |
    REST API for ShopQ — user app, vendor app, and admin panel.

    ## Authentication
    | App | Method | Header |
    |---|---|---|
    | User App | Bearer Token (Sanctum) | `Authorization: Bearer <token>` |
    | Vendor App | Bearer Token (Sanctum vendor guard) | `Authorization: Bearer <token>` |
    | Admin Panel | Bearer Token (admin guard) | `Authorization: Bearer <token>` |

    Tokens are returned from the respective login endpoints.
  version: "1.0.0"
  contact:
    name: ShopQ Dev Team
servers:
  - url: http://192.168.1.7:8000/api
    description: Local dev server
  - url: http://localhost:8000/api
    description: Localhost
...
components:
  securitySchemes:
    UserBearer:
      type: http
      scheme: bearer
      bearerFormat: Token
      description: "Sanctum token from POST /auth/login"
    VendorBearer:
      type: http
      scheme: bearer
      bearerFormat: Token
      description: "Sanctum token from POST /vendor/login"
    AdminBearer:
      type: http
      scheme: bearer
      bearerFormat: Token
      description: "Admin token from POST /admin/login"
...
paths:
...
  /products:
    get:
      tags: [Products]
      summary: Get all products (public)
      parameters:
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: limit
          schema: { type: integer }
      responses:
        "200":
          description: Product list
...
  /vendor/products:
    get:
      tags: [Vendor Products]
      summary: Get vendor's products
      security:
        - VendorBearer: []
      responses:
        "200":
          description: Product list
...